+++ src/panel/toolbox/VariableManager.vue
@@ -22,15 +22,20 @@ import Chat from '@/panel/toolbox/variable_manager/Chat.vue';
 import Global from '@/panel/toolbox/variable_manager/Global.vue';
 import Message from '@/panel/toolbox/variable_manager/Message.vue';
 import Preset from '@/panel/toolbox/variable_manager/Preset.vue';
+import { useCharacterSettingsStore } from '@/store/settings';
+
+const character = useCharacterSettingsStore();
 
 const active_tab = useLocalStorage<number>('TH-VariableManager:active_tab', 0);
 const tabs = [
   { name: t`全局`, component: Global },
   { name: t`预设`, component: Preset },
-  { name: t`角色`, component: Character },
-  { name: t`聊天`, component: Chat },
-  { name: t`消息楼层`, component: Message },
-];
+]
+  .concat(character.name ? [{ name: character.name, component: Character }] : [])
+  .concat([
+    { name: t`聊天`, component: Chat },
+    { name: t`消息楼层`, component: Message },
+  ]);
 </script>
 
 <style lang="scss" scoped>